Move App Module to module folder
